fix: Avoid recursive stack overflow in flake.nix
Dunno how I managed to reference `self` and why it's been working
someplaces and others not.

But in nix evaluation this will probably resolve to the store path of
the current flake. This will have already been added to the store. If
this reference then is picked up by nix's "trace inputs" mechanism then
this will lead to a derivation depending on itself, thus leading to an
infinite recursion.