deps: make pandas and numpy optional
We only use features of pandas when we get called with a pandas object
as a selector for either `cells` or `colors`. This is a reverse
catch-22, so dont we need to pull in pandas ourselves and rely on one
already being installed if required.

We still keep the `pandas-stubs` library for development.