dev: some fixes
- add makefile for easy running of pre-commit and bump-version
- ignore end-of-line errors for yaml files, this was fighting with my
  editor
- reformat the pre-commit config
- change the PR workflow to `pull_request_target`:
```
This event runs in the context of the base of the pull request, rather
than in the context of the merge commit, as the pull_request event does.
This prevents execution of unsafe code from the head of the pull request
that could alter your repository or steal any secrets you use in your
workflow. This event allows your workflow to do things like label or
comment on pull requests from forks. Avoid using this event if you need
to build or run code from the pull request.
```