vis: add interactive `log` property to python class
Similar to how `.x` and `.i` properties can be used interactively to
output ansi or image versions of a visual. This adds a log endpoint that
will print the latest logs.

The reason for making it a property that does a print instead of
returning the string is that 99% of the use case will be
`print(plt.log)` and it feels better to let the user just do `plt.log`
when exploring novem interactivly.